python-backend/requirements.txt
fastapi[all]==0.110.0
google-genai==0.1.0
python-multipart==0.0.9
pydub==0.25.1
python-dotenv==1.0.1
uvicorn[standard]==0.29.0